backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0; sys_platform != 'win32'
pymysql>=1.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-filter>=23.5
gTTS>=2.3.0
edge-tts>=7.0.0
deep-translator>=1.11.0
aksharamukha>=2.3
